feat(tests): update tests
